** Update - randomize user display name and user functionality to update display name **
Summary: This is kinda of a big update. When user signs in, we generate
a random display name for them. When they get into the feeds page. They
will see a banner showing their randomize display name and allows them
to update it to their desired display name

File Changes:
 - Splitting the settings/page.tsx
  - Keep the settings/page.tsx as a server component
  - Created a new component components/userSettingsForm.tsx
   - This is a client side component to handle a dynamic form submission
and validation
 - Update components/qoutesFeed.tsx
  - Adding the welcome banner for user who have not updated their
display name. User has option to go into the settings page or dismiss
the message
 - Update schemas/user.prisma to handle new fields for displayName
changes and updates. We used these fields for displayName and tracking
if the display is changed so we can show the banner in the feeds
component
 - Added new migration for the displayNameChange flag
 - Updated data/qoutes.ts: removed white spaces
 - Updated data/user.ts
  - added new return field for getUserByID
  - update updateUserDisplayName function to change the flag
displayNameChanged when the displayName is changed
  - added new function setDisplayNameDismissed so we can changed the
flag displayNameChanged when user clicked the dismiss button in the
banner
  - added new function getDisplayName to check if displayName already
exist while also checking that the displayName is not the current users
display name
 - Updated actions/user.ts
  - update the updateCurrentUserDisplayName function to handle updating
the display name for the user. We added multple checks and error
handling depending on the scenario of the display name changes
  - Updated auth.ts to set the displayNameChanged flag into the user
session so we can check it after logged in.
  - Updated next-auth.d.ts to handle the displayNameChanged on the
extended session object